+++ apps/meteor/ee/server/apps/communication/events.ts
@@ -1,0 +1,12 @@
+export enum AppEvents {
+	APP_ADDED = 'app/added',
+	APP_REMOVED = 'app/removed',
+	APP_UPDATED = 'app/updated',
+	APP_STATUS_CHANGE = 'app/statusUpdate',
+	APP_SETTING_UPDATED = 'app/settingUpdated',
+	COMMAND_ADDED = 'command/added',
+	COMMAND_DISABLED = 'command/disabled',
+	COMMAND_UPDATED = 'command/updated',
+	COMMAND_REMOVED = 'command/removed',
+	ACTIONS_CHANGED = 'actions/changed',
+}
