+++ apps/meteor/ee/server/apps/communication/websockets.ts
@@ -8,20 +8,9 @@ import { api } from '@rocket.chat/core-services';
 import { SystemLogger } from '../../../../server/lib/logger/system';
 import notifications from '../../../../app/notifications/server/lib/Notifications';
 import type { AppServerOrchestrator } from '../orchestrator';
+import { AppEvents } from './events';
 
-export enum AppEvents {
-	APP_ADDED = 'app/added',
-	APP_REMOVED = 'app/removed',
-	APP_UPDATED = 'app/updated',
-	APP_STATUS_CHANGE = 'app/statusUpdate',
-	APP_SETTING_UPDATED = 'app/settingUpdated',
-	COMMAND_ADDED = 'command/added',
-	COMMAND_DISABLED = 'command/disabled',
-	COMMAND_UPDATED = 'command/updated',
-	COMMAND_REMOVED = 'command/removed',
-	ACTIONS_CHANGED = 'actions/changed',
-}
-
+export { AppEvents };
 export class AppServerListener {
 	private orch: AppServerOrchestrator;
